[sestarnix] Remove conflicting gtest helper
The starnix test helper defines `SyscallSucceeds`.
The sestarnix test helper defines its own `SyscallSucceeds`,
with a slightly different API.

The 2 definitions conflict, preventing the use of the starnix
test helper in the sestarnix tests.

This CL removes the sestarnix version of `SyscallSucceeds`.
As a cleanup it also removes the sestarnix version of
`SyscallFailsWithErrno`.

Bug: None
Change-Id: Id07b4f766f3943b0bcda6c83022ff773e967e9ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1268265
Reviewed-by: Ambre Williams <ambre@google.com>
Commit-Queue: Jean-François Geyelin <jif@google.com>