[build][idk] Add CPU and API level args to idk_subbuilds()
Previously, `idk_subbuilds()` would use the global GN variables for CPU
architectures and API levels to support in the IDK. This may not always
be appropriate and obscures these details. Instead, add arguments to the
template and pass the global GN variables when appropriate.

Change-Id: Ibe6e06ec72fde4b913c23c363833b43a850e4ff9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1264946
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>