[starnix] Refactor UserAddress for overflow safety
This change refactors UserAddress and its API so that operations are
checked for overflow. Notably, we lose the ergonomics of += and -=
operations, since the trait definition doesn't allow for any memory-safe
way to modify a UserAddress in place and communicate potential faults
back to the caller. I think this is an acceptable loss.

Some notable changes in this commit include:

- Many of the syscalls return EFAULT now, which I've audited against the
  manpages for correctness. However, I did not audit every code branch
  and there may be places where we now EFAULT unexpectedly.
- msync transforms the EFAULT return code, based on the manpage for it.
- memory_manager now has some callsites which return early on erroneous
  address manipulation. This might cause side effects, but I'm not
  familiar enough with that code to deduce them.

Bug: 387826083
Test: TBD
Change-Id: I06079d626bfd491f30d0053aa372d46d00dfef2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1255695
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Aaron Knobloch <aknobloch@google.com>