[cpp][fidl] codegen cleanup
A very very common pattern when generating code for types that may
contain handles is to do:
  {{ if .IsResourceType }}
    {{ IfdefFuchsia }}
  {{ end }}
  // some stuff that can't build on host
  {{ if .IsResourceType }}
    {{ EndifFuchsia }}
  {{ end }}

I've added a pair of template functions so that this can be simplified
to:
  {{ IfdefFuchsiaIf .IsResourceType }}
  // some stuff that can't build on host
  {{ EndifFuchsiaIf .IsResourceType }}

The conditional is *always* .IsResourceType so this could have been
  {{ IfdefFuchsiaIfIsResourceType . }}
  // some stuff that can't build on host
  {{ EndifFuchsiaIfIsResourceType . }}
but I feel like that's a little bit too magic.

There are a couple of little other clean-ups in there that I came across
while working on it.

Change-Id: I649664f277505227fa7a7f3e777f951290250853
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1267024
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Koloski <dkoloski@google.com>