[power-manager] Remove fake dependency
It's not longer important for power manager to outlive driver manager.
As such we can safely remove the fake dependency which was added to
ensure it does outlive it.

The lifecycle stop hook was also implemented to ensure it can exit
gracefully once it needs to stop.

Fixed: 42071033
Test: component start power_manager; component stop power_manager
Change-Id: Ie606c1a5e5911de5ad9b57d4561c6212f48e2227
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1259367
Reviewed-by: Hai Bi <bihai@google.com>
Reviewed-by: Onath Dillinger <claridge@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>