[kernel][arm64] Remove secondaries_released synchronization
This change removes an unnecessary synchronization variable,
secondaries_released.  The variable is unnecessary because the
variable is set to true before the secondaries are even started.

Why is it safe to remove this?  The variable was set to true during
arch_init.  The secondaries are started by platform_init and would
spin until the value is true.  However, arch_init is called before
platform_init so by the time the secondary are started, it's already
true.

Bug: 414456459
Change-Id: I82229bd12487202a0af2a806cd2c054715fe0067
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1257704
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>