[trace_manager] Don't double copy buffer in streaming mode
The tracing system doesn't differentiate between setting
"rolling_data_end" to "0" to mean no data has been written, and "0", we
haven't yet committed the buffer. This is problematic in streaming mode,
as it causes us to attempt to uncondionally read the inactive buffer
after the trace ends. This happens because after we transfer the buffer,
we "reset" the rolling_data_end to "0".

In trace manager, when we copy the buffer out in streaming mode, we now
check last_wrapped_count_, which is our local copy of the last buffer we
transferred and compare it to the tracee's wrapped count before
attempting to copy the buffer.

Bug: 312790394
Change-Id: Ica680c5aeefbd3515a34d31720d276b33383c385
Multiply: fuchsia-pkg://fuchsia.com/trace_manager_integration_tests#meta/streaming_test.cm
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1266450
Reviewed-by: Jia Li <lizhijia@google.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>