[interconnect] Introduce a new core driver
This driver will be used as a layer between vendor interconnect drivers
and downstream clients. It takes a graph of nodes and edges and finds
the relevant paths between nodes that downstream clients will want to
set bandwidth parameters on. It aggregates requests from the various
clients and presents them to the vendor driver to take action on.

This driver was chosen to to be implemented in rust due to requirements
around heavy graph manipulation and the rust standard library generally
having an easier time of providing good tools for working with graphs
(short of using linked lists). Keeping the graph relatively flat rather
than a series of distributed linked lists was done to keep things
simpler.

Test: fx test interconnect
Bug: 405206028
Change-Id: Id03bc269084d71f0cdb6b21412cea228e9551eea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261365
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>