[kernel] Track dropped records in ktrace per-CPU mode
Modifies the per-CPU implementation of ktrace to track the duration,
number, and size of the records it drops due to buffers being full.

All tracking is done entirely by the writers, meaning that:
1. A writer begins tracking when it fails a reservation for the first
   time.
2. A writer emits a duration to the KTrace buffer containing the
   aforementioned statistics when there is space to do so.

An alternative design was considered in which the reader would perform
step 2 as soon as it read enough data out of the trace buffer, but this
was deemed more complex, as one would need to start thinking about how
to synchronize the reader's and writer's observation of the dropped
record statistics.

Bug: 404539312
Change-Id: I97819c3ef7d548abe0b7ba62d16956d020759ea5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1262925
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: John Grossman <johngro@google.com>