[storage][sparse] add resparse_sparse_img
resparse_sparse_img takes a SparseReader<Read + Seek>, an output
directory and a max_download_size and generates a set of Sparse Images
for it.

This departs from previous approaches which unsparse a sparse image to a
temporary file and then build a new sparse image from the unsparsed one
as it loses the resolution of what is a Skip Chunk (which is written as
a block of 0) and what is a Fill Chunk of 0.

Change-Id: I08d74563d429680688fc56d10a9b2c8a6a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1264027
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>
Reviewed-by: David Murphy <djmurphy@google.com>