[drivers][power] Refactor responsibilities
ManualWakeLease is now responsible for managing interaction
with SystemActivityGovernor protocols which includes:
  * Acquiring wake leases
  * Responding appropriately to suspend/resume notifications

ManualWakeLease no longer wraps a TimeoutWakeLease.

TimeoutWakeLease now wraps a ManualWakeLease.
TimeoutWakeLease is responsible for accepting a timeout
value and updating it as appropriate. A TimeoutWakeLease
uses ManualWakeLease's Start() and End() methods to
establish the time bounds during which we should claim a
wake lease, if the system is suspended, or claim a wake
lease if the system starts to suspend.

Bug: 412787822

Change-Id: I448f38be2085a1e77c9558ae40f95e0d2fd19198
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1264034
Fuchsia-Auto-Submit: Justin Mattson <jmatt@google.com>
Reviewed-by: Nick Eaton <nickeaton@google.com>
Commit-Queue: Justin Mattson <jmatt@google.com>