[dash] Add power builtin
This is a migration of commands currently living under dm. The dm
command stand for devmgr, the predecessor to driver-manager which used
to do a lot more in the system. It no longer makes sense to have the
power commands live underneath it.

Change-Id: Ib388e97e71dd66f6aca143340d2ec24cc67b2634
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1271608
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>