[ffx] Support "ffx -t target list"
A target list query can be specified with "ffx target list <query>", but
usually users specify a target with "-t". This CL allows "ffx -t <query>
target list" to also be used.

Note that the CL is slightly tricky, in that normally "-t" is handled
by querying the default target. But we don't want the _default_ target
to override, we only want to override if "-t" was specified on the
command line.  So we explicitly look at the FhoEnvironment copy of the
FfxCommandLine to implement the correct logic.

This introduces a behavior change: if a script were already calling
"ffx -t <target> target list" and checking for a exit status of 0,
this will break such functions. Such code should be changed to not
specify a target unless it is actually checking for the specified
target.

Bug: 344903757
Change-Id: I02d12c5a3b2fbacfa83410f48452541bd150c029
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1269066
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>