[sestarnix] Fail component launches without Security Context
If the Starnix container has SELinux configured then either the
container or the component must specify the Security Context under
which it should run. This removes the fall-back to inheriting the
Security Context of the init process, which is invariably not what
the component author expects.

Change-Id: I5781539d188664818c134cfc088f7ec0a222e5d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1271005
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>