[fx] suggest correct commands when there is a typo
This changes adds a script which will run when there are no commands
found when executing fx. The script will use an edit distance to figure
out if the user had a typo in their command and suggest the correct
command. If there are not suggestions it will fall back to the original
error message.

With this change typos will now be reporting as:
```
$ fx buidl
ERROR: Command 'buidl' not found. Did you mean 'build'?

$ fx emu
ERROR: 'fx emu' is no longer valid, use 'ffx emu' instead
```

Bug: 402762037
Change-Id: I31e7e423a02b804aeae17a7eeaf66d0f10f9ae05
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1271284
Reviewed-by: David Murphy <djmurphy@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>