[kernel] Return required buffer size from zx_ktrace_read
The existing implementation of zx_ktrace_read returns the amount of
available data when the passed in user pointer is nullptr. This change
modifies the per-CPU implementation to match this behavior, with the
slight difference that it always returns the total size of all per-CPU
buffers, since that's what we require in this mode.

Bug: 404539312
Change-Id: Ie2887c5f3492ef4bfff26a218b81914b01586da8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272209
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: John Grossman <johngro@google.com>