[kernel] Make legacy ktrace test a no-op when streaming is enabled
The TestKTrace::LegacyCategories unit test verified that categories
worked correctly when using the single buffer implementation. This
test is no longer useful when streaming is enabled, so add an ifdef
that disables it in that case.

Once we've switched to streaming everywhere, we can delete this test.

Bug: 404539312
Change-Id: I14bd7f7bc33012ff1dfea109b993989b1745dc68
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1271875
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>