[kernel][vm] Prevent NextChild returning root_
Corrects the TreeWalkCursor::NextChild method to avoid spuriously
returning the root_ node, specifically in the VmCowPages::Resize
operation (this spurious return is benign when performing a
RangeChangeUpdate).

In a race between dropping the current lock and acquiring the child
lock the child could be destroyed, causing it to not have a parent_. In
this case we, correctly, call UpdateCurLocked to re-establish cur_,
which is still the root_, but then just return despite not having
actually moved to a child.

This commit corrects this and will retry internally in NextChild if a
race causes the node deletion so that the stated contract of NextChild
is upheld (that it moves to a child or sibling or returns false).

Although not the cause of the specific assertion failure in b/415665459
this was found during that investigation as it causes other, non
kernel assertion triggering, errors in the Resize operation.

Bug: 415665459
Change-Id: Id524a9255e3f0f06697ace605ab8c597d5816903
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1271665
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>