[component_manager] add FIDL to get own sandbox
Adds a new FIDL API for a framework protocol that can be used to get a
component's own sandbox.

As this is not yet ready for wide consumption, the implementation of the
protocol checks that the component accessing the protocol is a builtin
component. In other words, component manager is currently the only
allowed consumer of this protocol.

Fixed: 415968285
Change-Id: Icfae9fc9740256dc615b4e3b8bba18ab9ea18e70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1269704
API-Review: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Olivia Nordquist <nolivia@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>