[component_manager] add built-in dispatcher component
The dispatcher component can be given configuration capabilities telling
it what capability name and type its dispatching, and the URL of a
component to dispatch to. It will expose a dynamically populated
dictionary holding the capability it dispatches. Any routing requests to
that capability will cause the dispatcher to launch a new worker
component with the given URL, and it will then forward the route request
to the worker. The worker is watched for a stop event, at which point
the worker is destroyed.

Any capabilities offered to the dispatcher are passed through to the
workers.

Fixed: 415968290
Change-Id: I1060025c0ab5d0f681a852786fcfc5e27c004bbd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1245065
Fuchsia-Auto-Submit: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Amanda Tait <atait@google.com>