[fshost] Remount the system Starnix volume on mount if mounted
Right now, fshost will fail StarnixVolumeProvider.Mount if the Starnix
volume is already mounted (AKA if Starnix failed to unmount its volume
on shutdown). In an effort to make fshost resilient to unclean shutdowns
of the Starnix container, fshost will now unmount and remount the
Starnix volume on StarnixVolumeProvider.Mount if it sees that the volume
is already mounted.

Bug: b/416031021
Change-Id: I5cb8d706491d0413ab54bf3b1d715f4c29491eec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272007
Commit-Queue: Nikita Jindal <nikitajindal@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>