[ffx] Add "is_manual: bool" to ffx TargetInfo
We'd like to add an indication to users of whether a target as
added manually or not when using ffx target list. This CL implements
one part of the puzzle -- allowing the daemon to provide whether it
is manual when a client queries the TargetInfo.

An alternative, full-featured data structure was also a possibility
(indicating the source -- discovered by mDNS, USB, etc), but the existing
daemon code is so convoluted that it would be difficult to add only
to achieve a simple goal, and since the daemon will eventually be
going away it didn't seem like a good time investment.

Bug: 287693356

Change-Id: I6a6a69641ac5e831a398288deef637c2112dc19d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1271886
API-Review: Chase Latta <chaselatta@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>