[netstack3] Hold scope guard for ReferenceReceiver
ReferenceNotifier panics if the corresponding receiver is dropped when
we're trying to notify resource removal. This is a useful mechanism to
catch misuse of the resource removal API, but that means dropping
receivers on scope cancelation during shutdown can trigger the panic.

Keep the panic in the notifiers side by ensuring the receiver is not
dropped due to the future being dropped when spawned in a scope with a
ScopeActiveGuard. Fallback to the deferred removal mechanism when we
can't acquire a guard.

Split the receiver and notifier implementations into their own module so
we can keep this neatly separated from the rest of bindings and test in
isolation.

Fixed: 416692381
Multiply: dns-integration-test

Change-Id: I14e44f435b016bb4bf71169ede6a17e8e98408bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272439
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>