[starnix][binder] Fix transaction dead replies
Previously we would insert `DeadReply`s into the sending thread's
command queue regardless of whether or not the reply was to the
currently active transaction of the sending thread.

This meant that the sending thread would process the `DeadReply` out of
order, and pop the wrong transaction off the stack.

This was an uncommon scenario, as it requires both a cycle in the binder
transactions between processes, and for one of the processes in the
cycle to die while multiple transactions are in flight.

Bug: 414799731

Change-Id: I49b1795ea9bf5478bd8a60c62e3d81e75b036a0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1273164
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Barth <abarth@google.com>