[kernel] Move per-CPU ktrace metadata record emission to boot CPU
Prior to this CL, the trace records containing the interned strings and
CPU koids were emitted on whatever CPU started tracing, not the boot
CPU. This led to situations in which strings were referenced before
the corresponding metadata record was seen. This worked in perfetto,
which can handle out of order records like this, but did not work
in the trace-reader userspace library, as that library needs to be
able to perform chunked reads. This CL fixes the issue by always
forcing the string refs to be emitted on the boot CPU.

Note that thread and process names are still emitted on whatever CPU
invokes ReportMetadata. This is fine because those records do not
need to be parsed before any other trace records.

Bug: 404539312
Change-Id: I284df5795048b17e2277712db3e1d0a3939d5c7f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272990
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>