[ffx] capture target list info in analytics
In order to understand the number of devices people generally have,
we'd like to add metrics for the number of devices, as well as the
type of query (not the query itself since that could be PII).

The event is called "ffx_target_list_devices", and it has two
pieces of information: the type of query ("first", "serial", etc),
and the number of devices returned.

Fixed: 334926773

Change-Id: I6a6a6964278a2ccf0cb2da30e7e076b2e202a26e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272820
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>