[sdk/test_manager] add break_on_failure
This CL adds the break_on_failure option to RunSuiteOptions. It already
appears in RunOptions, which is deprecated.

This brings the new test_manager API up-to-date with respect to the
deprecated API and is part of the effort to remove the old API.

Bug: 309995479
Test: no behavior change, FIDL definition only

Change-Id: I29891be2347be11f68c213bc889ba5e8c5bb558f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272210
API-Review: Dale Sather <dalesat@google.com>
Fuchsia-Auto-Submit: Dale Sather <dalesat@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>