[starnix] Add default domainname for proc
The splice06 LTP test does not expect the domainname file to return a
zero-byte value. I think that's technically a bug, and I've started a
convo with upstream maintainers. Nonetheless, on my gLinux workstation
and laptop, even though a domainname is not configured, cat'ing the file
returns the string `(none)`. We should emulate that in order to appease
the Android VTS / LTP test suite.

Bug: 409149432
Test: fx test starnix_ltp_tests_s -o --test-filter "splice06"
Change-Id: Ib5e5c297a424b9dd37516886632d5942b715885c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1273386
Reviewed-by: Vickie Cheng <vickiecheng@google.com>
Commit-Queue: Aaron Knobloch <aknobloch@google.com>