[build][idk] Remove api_contents defined check
After https://fxrev.dev/1270044, `api_contents` is the only way to
specify the API contents, so remove a check for whether it is defined.

Also removed an unnecessary amending of visibility to allow the source
set atom to access the underlying source set because it does not.

Bug: 412696706
Change-Id: Id3097ae5a8c1cdf03327ad2174c15ef8f2b2c4d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1273892
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>
Reviewed-by: David Turner <digit@google.com>