[starnix] Add support for effective SEStarnix SID
In some situations (communication with external subjects, Overlay FS)
the Starnix kernel needs to act on behalf of a task that is not the
current task. This change introduces an effective SID in the SEStarnix
task state. When doing access checks where the current task is the
subject, this effective SID is used instead of the current SID.

To avoid security issues, the effective SID can only be used from the
current task: anything acting on the task as an object, and ptrace
checks on exec that are done with the context of another task, are
performed with the current SID instead. We enforce this by requiring a
CurrentTask to access the effective SID. This requires moving some
operations doing capability checking from Task to CurrentTask: in
practice they were only ever called on the current task anyway.

Some operations don't make sense in contexts where the current and
effective SIDs would be distinct (e.g. "exec" should be called by
the process itself, not on behalf of another subject). We assert
in these cases that the SIDs are distinct: a failure here indicates
a Starnix issue, not an application issue.

This change shouldn't change any access checking, as nothing sets an
effective SID that is distinct from the current SID yet.

Bug: 413607595
Change-Id: I1259cf1c4f04e0125cb9a703b8347fa0ed0a16d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1261404
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Ambre Williams <ambre@google.com>