[memory] Add memory stalls to regular measurements
Add memory stalls to the measurements shown in Inspect by Memory Monitor
2. This will allow to display the stalls alongside bucket memory values,
if we so wish.

This change also removes the display of the stall rate, as it can now be
computed directly from the Inspect output. This paves the way to
repurpose the stall rate computation for Cobalt metrics.

BUG=399440175

Change-Id: I95c51f99d43fa363a4ed4633c820f7d259059444
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1274026
Commit-Queue: Étienne J. Membrives <etiennej@google.com>
Reviewed-by: Guillaume Chelfi <chelfi@google.com>