[ftl] Disable new wear leveling algorithm
Partial revert of commit 289f67706fcedf67a7ba073215f900e0451841aa.

This separates the driver controls that are relied upon by out of tree
callers and the actual implementation. It allows the external caller to
still set the bit and responds the same way, but will log a warning that
the setting will not actually be honoured.

Bug: 416695119
Change-Id: Id006c499ceb3f77efbb114b31c4c426eb0e03b0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1275404
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>