[wlan][policy] Logs disconnect on channel closure.
Prior to this change, the telemetry module was not notified
of disconnect when a client SME event stream terminated.
When driver recovery was the cause of this breakage, the
telemetry module would repeatedly attempt to query
statistics for a non-existent interface.

This change causes the client state machine to notify the
telemetry module that the client has disconnected on exit
which results in the telemetry module transitioning from the
Connected state into the Idle state so that interface
statistics will no longer be queried.

Test: fx test -o wlancfg-tests
Multiply: wlancfg-tests
Change-Id: Ie2b8a2bb4830cb0a910ea02f7ad74e4695a2fa8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272124
Commit-Queue: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>