[starnix][procfs] Untangle TaskDirectory from CurrentTask
Previously, the CurrentTask had knowledge of the TaskDirectory in order
to change the uid/gid for the /proc/pid directory when the credentials
of the current task changed. Now, we propagate this information without
needing knowledge of the TaskDirectory type.

To achieve this effect, we no longer cache the FsCreds object for the
TaskDirectory in the TaskDirectory struct. Instead, we recover the
FsCreds either from the underlying TaskObject itself or from the
FsNodeInfo object associated with the directory.

Change-Id: I64bda73f7425c779b569dba9983e453d693a97bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1275444
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>