[cpu_profiler] Better encapsulate component attaching
The logic to walk a component and attach to it leaked out into the
profiler_controller code. It also meant that the ComponentTargets were
having callbacks injected into them which is difficult to follow for a
reading, and also copies raw `this` pointers which is spooky.

This should help make the job to pregather all modules in the sampler
easier to reason about.

This:
- pulls the component walking logic into component.[cc,h]
- switches to using a Ref+Weak pointer so that we don't have dangling
  `this` references hanging around
- Changes component to only have access to the sampler instead of saving
  a `this` to the profiler_controller global state.

Bug: 417233162

Change-Id: Icb27ba3c23269b2e03edac7ffe98b87c0f5e5ea8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1274745
Reviewed-by: Jia Li <lizhijia@google.com>
Reviewed-by: Mary Xia <maryxia@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>