[cpu_profiler] Hold Process Modules in a Map
We'll be grabbing process modules and the beginning of the session, and
we'll want to update them at the end to grab any possibly new modules.
Switch to a map data structure keyed by build id so that it's easier to
incrementally update the newly found modules.

Bug: 417233162

Change-Id: I3ac2cf4f1ef033234e49ac1ef383473484d0b9e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1274746
Reviewed-by: Mary Xia <maryxia@google.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Jia Li <lizhijia@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>