[bt][hfp-hf] SCO setup and code negotiation.
This commit adds support to to the peer task for awaiting incoming SCO
connections and SCO connect closes.  As part of this, it starts and stop
A2DP audio as needed. Some common functionality for SCO is moved to a
common library and some some changes are made to the AG implementation
as well.

Some other small things are in this commit as well, namely using the
strongly typed CodecId type and changing the name of
ProcedureManagedState to ProcedureManipulatedState to avoid confusion
with the ProcedureManager.

Bug: b/42061084
Test: Tests in a coming commit.
Change-Id: I82339de1eda735d051e87c67ba127d285d437bfc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1257844
Commit-Queue: Matthew Kehrt <kehrt@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>