[usb-peripheral] Migrate to DFv2
Migrate the usb-peripheral driver to DFv2.

Make the usb-peripheral store UsbFunction instances in a vector instead
of fbl pointers. The goal is to make it clear that the driver owns the
UsbFunction instances and more easily catch uses after the driver has
stopped.

Make the driver clear usb-functions by simply removing them instead of
waiting for the usb-functions to signal to the driver that they have
been removed. Originally, this was necessary because the usb-functions
had to asynchronously remove the DFv1 device that they owned. This is no
longer the case as they instead own FIDL clients to DFv2 nodes which can
be dropped immediately.

Bug: 416554395
Change-Id: I12ef1c41b196fada5f4ab6d8b622a7a885a1b576
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1266884
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>