[kernel] Fix ktrace per-CPU buffer init with uneven buffer size
The previous implementation assumed that the provided target buffer
size will divide cleanly among the number of CPUs in the system,
and that the resulting buffer size will be a power of two. This is
obviously not always the case.

This CL fixes the issue by dividing the given buffer size by the
number of CPUs, and then rounding the resulting value down to the
nearest power of two.

Bug: 404539312
Change-Id: Ibb5fadbb3ee723bc3b87953c0b631b96484de86e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1276860
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: John Grossman <johngro@google.com>