[kernel][vm] Introduce ParentContent markers
Adds a new type of marker, the ParentContent marker, that are used in
non-hidden nodes that do not have a root page source. The purpose of
the ParentContent marker is to allow for eliding a walk to the root in
the case where there is no content.

The insertion and cleanup of these markers adds significant overhead to
the VMO clone and close steps of micro benchmarks, but leveraging
ParentContent to elide walk ups makes up for it.

Bug: 338300943
Change-Id: I5a95e5289d7d6ab40e96591ac8e0c38ad26266e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1254746
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>