[power] Increase history capture for PB and SAG
This patch increases the number of events captured
by Power Broker and System Activity Governor, to
improve diagnosability beyond ~5-10 minutes of history
(in the worst case we see in the fleet today).

With recent fuchsia.inspect.Graph optimizations, PB's
increase to 16k events occupies 82% of the VMO, which
leaves some headroom for safety and future adjustments.

SAG's increase to 4k events requires extra room
room in the VMO, and we see a steady state occupancy
of 85%. (For the "1000 leases" test, we see occupancy
of 93%.)

Bug: 394674201
Change-Id: I00f2089be235bc24def115227c50b93cdb3e8f14
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1270428
Reviewed-by: Onath Dillinger <claridge@google.com>
Commit-Queue: Jaeheon Yi <jaeheon@google.com>