[kernel][vm] Stack canary
Adds a stack canary, controllable via the kernel cmdline, that writes a
small canary pattern near the end of a stack to, heuristically,
determine if a high watermark of stack usage was exceeded.

Detection is performed on stack / thread deletion and so will not find
excessive stack usage on forever executing threads, or stack usage that
allocated, but did not write to, the portion of the stack where the
canary was. This limitations can be improved upon in future CLs.

Bug: 417302663
Change-Id: Icdf02200bfc14937bac45eeba2bb27d88fcf068f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272605
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>