[spmi] Add dump support to spmi-ctl
Add a dump command that reads 1 byte at the time and does not
stop on errors.
This allows a user to read large amounts of registers even if
some would fail the command due to an error for some of the
registers.

Also make read length a hex input, updating the help to match.

Test: spmi-ctl reads on target
$ spmi-ctl -t 0 -a 0x9000 -d 0x7000
Executing on controller: msm-spmi
Register: 0x9000  value: 0x02 (2)
...

Change-Id: If26994b64e141574e14e85f427bb2d371d2824b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1268664
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>