[owners] Remove miguelfrde@google.com from tools/devshell/contrib/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/miguelfrde@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I49e980449151c866c7279469abb9ae82077a4d39
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1270025
Fuchsia-Auto-Submit: Owners-Cleanup-Prod <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Andrew Davies <awdavies@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit-builder@fuchsia-internal-service-accts.iam.gserviceaccount.com>
Reviewed-by: Andrew Davies <awdavies@google.com>