[netcfg/dns] Socketproxy netemul tests
Add tests to validate the new DNS server source for socketproxy-provided
DNS servers. Socketproxy DNS servers will be provided from the
DnsServerWatcher when there is no Fuchsia-provisioned default network
set. In these instances, the DNS servers will reflect the set of servers
known by the Starnix Networks currently in the StarnixNetworks registry.

The implementation of this functionality is in https://fxrev.dev/1271084
and is tested in this CL for ease of review.

Fixed: 392152278
Change-Id: Ia5ca6277f0f67b333d6094b3f28ffb67eff1b9eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192691
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>