[sestarnix] Hooks to temporarily patch task attributes
This change introduces hooks to temporarily patch task attributes, and
restore them after performing an operation. Internally, the changes to be
performed are described by a `TaskAttrsPatch` structure, composed from
provided operations. Currently, two operations are exposed: using an
effective SID previously copied from a CurrentTask, and copying the
security state of an FsNode as an fscreate SID. These are exposed to
the outside-SEStarnix world as two independent hooks.

To ensure that the state is properly restored, we require callers to
provide a closure that we wrap in the apply/restore operations. In
contrast to providing an object with a `Drop` implementation that
performs the restoring, this ensures that the lifetime of the
temporary state cannot be mistakenly extended, by moving or
leaking. In particular, the state will be restored before returning to
user code.

Bug: 413607595, 364568696, 369067922, 364568931
Change-Id: I565dfa32f25706a2ef373a1bf306195defc5876b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1275225
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Ambre Williams <ambre@google.com>